cloud-computing/requirements.txt
Flask
Werkzeug
Pillow
Keras
numpy
PyJWT
keras-preprocessing
python-dotenv
Flask-SQLAlchemy
PyMySQL
tensorflow
mysqlclient